feat: terminando

+++ p1_ldan/app/src/main/java/com/example/p1_ldan/MainActivity.java
@@ -29,7 +29,7 @@ public class MainActivity extends Activity {
 
                 double valorDesejado = Double.parseDouble(iValorDesejado.getText().toString());
                 double taxa = Double.parseDouble(iTaxa.getText().toString());
-                double periodo = Integer.parseInt(iPeriodo.getText().toString());
+                int periodo = Integer.parseInt(iPeriodo.getText().toString());
 
                 Intent resultCDI = new Intent(MainActivity.this, ResultActivity.class);
                 resultCDI.putExtra("taxa", taxa);
